Convert parallel tests to mocha
